Use repo url helper
